ACS-6256: Move logic related with managing folders and categories for TAS tests from acs-packaging to community-repo - addressing PR comments
