+++ packaging/tests/tas-restapi/src/main/java/org/alfresco/rest/repo/resource/ResourceManager.java
@@ -37,7 +37,6 @@ import org.alfresco.utility.model.TestModel;
  * @param <SPECIFIER> repository resource specifier, see {@link Specifier}
  * @param <MODIFIER>> repository resource modifier, see {@link Modifier}
  */
-@SuppressWarnings({"PMD.GenericsNaming"})
 public interface ResourceManager<RESOURCE extends TestModel, SPECIFIER extends Specifier, MODIFIER extends Modifier<RESOURCE, ?>>
     extends ResourceIntroducer<SPECIFIER>
 {
@@ -51,4 +50,9 @@ public interface ResourceManager<RESOURCE extends TestModel, SPECIFIER extends S
     }
 
     void delete(RESOURCE resource);
+
+    default void delete(String id)
+    {
+        delete(get(id));
+    }
 }
