+++ packaging/tests/tas-restapi/src/main/java/org/alfresco/rest/repo/resource/general/ContentModifier.java
@@ -43,7 +43,6 @@ import org.alfresco.utility.model.ContentModel;
 import org.alfresco.utility.model.FolderModel;
 import org.alfresco.utility.model.SiteModel;
 
-@SuppressWarnings({"PMD.GenericsNaming"})
 public abstract class ContentModifier<CONTENT extends ContentModel, SELF extends Modifier.ContentModifier<CONTENT, ?>>
     extends ResourceModifier<CONTENT, SELF>
     implements Modifier.ContentModifier<CONTENT, SELF>
