+++ packaging/tests/tas-restapi/src/main/java/org/alfresco/rest/repo/resource/general/Modifier.java
@@ -39,7 +39,6 @@ import org.alfresco.utility.model.UserModel;
  * @param <RESOURCE> repository resource, e.g. folder, category, etc.
  * @param <SELF> return type - this interface extension or implementation
  */
-@SuppressWarnings({"PMD.GenericsNaming"})
 public interface Modifier<RESOURCE extends TestModel, SELF extends Modifier<RESOURCE, ?>>
 {
     RESOURCE get(String id);
