+++ packaging/tests/tas-restapi/src/main/java/org/alfresco/rest/repo/resource/general/MultipleContentsCreator.java
@@ -38,7 +38,6 @@ import org.alfresco.utility.model.UserModel;
 import org.apache.commons.collections.CollectionUtils;
 import org.apache.commons.lang3.RandomStringUtils;
 
-@SuppressWarnings({"PMD.GenericsNaming"})
 public abstract class MultipleContentsCreator<CONTENT extends ContentModel, SELF extends MultiCreator.ContentsCreator<CONTENT, ?>>
     extends MultipleResourcesCreator<CONTENT, SELF>
     implements MultiCreator.ContentsCreator<CONTENT, SELF>
