+++ packaging/tests/tas-restapi/src/main/java/org/alfresco/rest/repo/resource/general/ResourceIntroducer.java
@@ -30,7 +30,6 @@ package org.alfresco.rest.repo.resource.general;
  *
  * @param <SPECIFIER> repository resource specifier, see {@link Specifier}
  */
-@SuppressWarnings({"PMD.GenericsNaming"})
 public interface ResourceIntroducer<SPECIFIER extends Specifier>
 {
     SPECIFIER add();
